requirements.txt
wheel
pydantic
pydantic-settings
Pillow
Scrapy